# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/registrations": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPatientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_registerAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPatientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "post": {
        "operationId": "AuthController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_endAllSessions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndAllSessionsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SessionResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "DermaHealth Backend API",
    "description": "Clinic workflow backend: identity, scheduling, clinical, workflow, EMR, CRM.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "UserMembershipResponseDto": {
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "clinicLocationId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "departmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "super_administrator",
              "patient",
              "doctor",
              "nurse",
              "receptionist",
              "lab_technician",
              "imaging_technician",
              "pharmacist",
              "care_coordinator",
              "customer_care_employee",
              "medical_administrator",
              "system_administrator",
              "clinical_process_designer"
            ]
          }
        },
        "required": [
          "organizationId",
          "clinicLocationId",
          "departmentId",
          "role"
        ]
      },
      "CurrentUserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "description": "Short-lived or transformed URL; null until T09 file service ships."
          },
          "status": {
            "type": "string",
            "enum": [
              "invited",
              "active",
              "locked",
              "disabled"
            ]
          },
          "activeOrganizationId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserMembershipResponseDto"
            }
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phone",
          "avatarUrl",
          "status",
          "activeOrganizationId",
          "memberships",
          "version"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/CurrentUserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "user"
        ]
      },
      "RegisterPatientRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "Minimum 12 characters."
          },
          "name": {
            "type": "string"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "1995-03-15"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "unknown"
            ]
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "organizationId": {
            "type": "string",
            "format": "uuid",
            "description": "Organization id when known."
          },
          "organizationCode": {
            "type": "string",
            "description": "Public organization code, e.g. dermahealth."
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "dob",
          "gender",
          "phone"
        ]
      },
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "nguyenvana@example.test"
          },
          "password": {
            "type": "string",
            "example": "a-strong-password"
          },
          "rememberMe": {
            "type": "boolean",
            "default": false
          },
          "mfaCode": {
            "type": "string",
            "description": "Reserved for MFA rollout; not yet enforced."
          }
        },
        "required": [
          "email",
          "password",
          "rememberMe"
        ]
      },
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": {}
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {}
      },
      "EndAllSessionsRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "Optional re-confirmation of the current password."
          }
        }
      },
...
    }
  }
}